Perform "git obliterate" on "content/" folder
Originally this repository was ment to keep track of changes on the
content of the website. Then it evolved in encompass the changes in
the building system of the website (which eventually became Zachaire).
However, it is a bad idea to publish a tool (Zachaire) that has
personal data (whatever was on my website) in the git repository and
history. The content folder should hence not be part of the main git
repository, but it may host another git repository that tracks the
changes made to the content/ directory and is meant to be kept private.
The tree structure is hence:

    myWebsite/
    ├── .git/     # Tracks changes in Zachaire,
    │             # is a public repository (e.g. may be pushed to github)
    ├── build.py
    ├── content/
    │   ├── .git/ # Tracks changes in the content of the website,
    │   │         # is a private repository (e.g. won't get pushed to github)
    │   └── index.html
    ├── LICENSE
    ├── Makefile
    ├── out/
    └── theme/

Note: for that reason, the labels of the commits preceding this one
might refer to the "content/" folder and are hence not representative
of the changes visible in the git history in its current state.